@@ -9,4 +9,4 @@ require (
 	gopkg.in/yaml.v2 v2.2.1
 )
 
-go 1.15
+go 1.16
