@@ -5,15 +5,15 @@
 package main
 
 import (
-	"io/ioutil"
+	"io"
 	"log"
 	"os"
 )
 
 // Loggers
 var (
-	Trace = log.New(ioutil.Discard, "[TRCE]", log.LstdFlags|log.Lmicroseconds|log.Lshortfile)
-	Debug = log.New(ioutil.Discard, "[DBUG]", log.LstdFlags|log.Lshortfile)
+	Trace = log.New(io.Discard, "[TRCE]", log.LstdFlags|log.Lmicroseconds|log.Lshortfile)
+	Debug = log.New(io.Discard, "[DBUG]", log.LstdFlags|log.Lshortfile)
 	Info  = log.New(os.Stdout, "[INFO]", log.LstdFlags)
 	Warn  = log.New(os.Stderr, "[WARN]", log.LstdFlags|log.Lshortfile)
 	Error = log.New(os.Stderr, "[ERRO]", log.LstdFlags|log.Lshortfile)
